requirements.txt
absl-py==2.1.0
# numpy==2.0.1
scipy==1.14.0
wget==3.2
opencv-python==4.10.0.84
tqdm==4.66.4
attrs==24.2.0
pillow==10.4.0
vertexai
transformers==4.42.4